+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add system mailer test functionality
 - Clean up inactive worker PIDs
 - Add query optimizer to MongoLite
+- Make IndexLite more compatible with Meilisearch
 
 ## 2.12.1 (2025-10-26)